+++ macx/macx.pro
@@ -1,6 +1,6 @@
 TEMPLATE = subdirs
 
 CONFIG += debug_and_release
-SUBDIRS = overlay compat osax
+SUBDIRS = compat osax
 
 include(scripts/scripts.pro)
