+++ overlay_gl/avail_mac.h
@@ -1,0 +1,52 @@
+#define AVAIL_ALL_GLSYM ( \
+	AVAIL(glPopAttrib) && \
+	AVAIL(glDisableClientState) && \
+	AVAIL(glEnableClientState) && \
+	AVAIL(glCreateShader) && \
+	AVAIL(glTexEnvi) && \
+	AVAIL(glTexImage2D) && \
+	AVAIL(glLinkProgram) && \
+	AVAIL(glVertexPointer) && \
+	AVAIL(glMatrixMode) && \
+	AVAIL(glGetString) && \
+	AVAIL(glBlendFunc) && \
+	AVAIL(glPopMatrix) && \
+	AVAIL(glDrawArrays) && \
+	AVAIL(glViewport) && \
+	AVAIL(glEnableVertexAttribArray) && \
+	AVAIL(glDeleteTextures) && \
+	AVAIL(glOrtho) && \
+	AVAIL(glBindTexture) && \
+	AVAIL(glColorMaterial) && \
+	AVAIL(glTexParameteri) && \
+	AVAIL(glPushAttrib) && \
+	AVAIL(glTexSubImage2D) && \
+	AVAIL(glCompileShader) && \
+	AVAIL(glShaderSource) && \
+	AVAIL(glGetIntegerv) && \
+	AVAIL(glDisable) && \
+	AVAIL(glPushMatrix) && \
+	AVAIL(glGetError) && \
+	AVAIL(glGetUniformLocation) && \
+	AVAIL(glAttachShader) && \
+	AVAIL(glGetBooleanv) && \
+	AVAIL(glPopClientAttrib) && \
+	AVAIL(glLoadIdentity) && \
+	AVAIL(glUniform1i) && \
+	AVAIL(glGetTexParameterfv) && \
+	AVAIL(glGenTextures) && \
+	AVAIL(glUseProgram) && \
+	AVAIL(glGetShaderInfoLog) && \
+	AVAIL(glTexParameterfv) && \
+	AVAIL(glBindBuffer) && \
+	AVAIL(glIsTexture) && \
+	AVAIL(glPixelStorei) && \
+	AVAIL(glActiveTexture) && \
+	AVAIL(glPushClientAttrib) && \
+	AVAIL(glDisableVertexAttribArray) && \
+	AVAIL(glRenderMode) && \
+	AVAIL(glGetVertexAttribiv) && \
+	AVAIL(glCreateProgram) && \
+	AVAIL(glEnable) && \
+	AVAIL(glTexCoordPointer) && \
+	1)
