+++ overlay_gl/avail_mac.pl
@@ -1,10 +1,10 @@
 #!/usr/bin/perl
-# perl avail.pl < overlay.m > avail.h
+# perl avail_mac.pl < overlay.c init_mac.c > avail_mac.h
 
 %funcs = ();
 while (<STDIN>) {
 	foreach (split(/([ \t])/, $_)) {
-		@glfunc = ($_ =~ /(gl[A-Z0-9][a-zA-Z0-9]+)/);
+		@glfunc = ($_ =~ /(gl[A-VY-Z0-9][a-zA-Z0-9]+)/);
 		foreach $func(@glfunc) {
 			$funcs{$func} = 1;
 		}
